+++ app/icons/icon-192.png/route.tsx
@@ -1,5 +1,0 @@
-import { renderIcon } from "@/lib/utils/ogIcon";
-
-export function GET() {
-  return renderIcon(192);
-}